+++ common/src/main/java/com/verdantartifice/primalmagick/common/wands/WandCap.java
@@ -7,6 +7,7 @@ import net.minecraft.network.codec.ByteBufCodecs;
 import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.resources.Identifier;
 import net.minecraft.world.item.Rarity;
+import org.jetbrains.annotations.NotNull;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
@@ -63,16 +64,19 @@ public class WandCap implements IWandComponent {
     }
     
     @Override
+    @NotNull
     public String getTag() {
         return this.tag;
     }
     
     @Override
+    @NotNull
     public Rarity getRarity() {
         return this.rarity;
     }
     
     @Override
+    @NotNull
     public Type getComponentType() {
         return IWandComponent.Type.CAP;
     }