+++ neoforge/src/main/java/com/verdantartifice/primalmagick/common/items/wands/ModularStaffItemNeoforge.java
@@ -1,9 +1,9 @@
 package com.verdantartifice.primalmagick.common.items.wands;
 
-import com.verdantartifice.primalmagick.common.wands.IWandComponent;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.context.UseOnContext;
+import org.jetbrains.annotations.NotNull;
 
 public class ModularStaffItemNeoforge extends ModularStaffItem {
     public ModularStaffItemNeoforge(Properties properties) {
@@ -11,24 +11,19 @@ public class ModularStaffItemNeoforge extends ModularStaffItem {
     }
 
     @Override
-    public InteractionResult onItemUseFirst(ItemStack stack, UseOnContext context) {
+    @NotNull
+    public InteractionResult onItemUseFirst(@NotNull ItemStack stack, @NotNull UseOnContext context) {
         return this.onWandUseFirst(stack, context);
     }
 
     @Override
-    public boolean shouldCauseReequipAnimation(ItemStack oldStack, ItemStack newStack, boolean slotChanged) {
+    public boolean shouldCauseReequipAnimation(@NotNull ItemStack oldStack, @NotNull ItemStack newStack, boolean slotChanged) {
         return false;
     }
 
     @Override
-    public boolean canContinueUsing(ItemStack oldStack, ItemStack newStack) {
+    public boolean canContinueUsing(@NotNull ItemStack oldStack, @NotNull ItemStack newStack) {
         // Don't break wand interaction just because the stack NBT changes
         return true;
-    }
-
-    @Override
-    public int getEnchantmentValue(ItemStack stack) {
-        // The enchantability of a staff is determined by its components
-        return this.getComponents(stack).stream().mapToInt(IWandComponent::getEnchantability).sum();
     }
 }