+++ neoforge/src/main/java/com/verdantartifice/primalmagick/common/items/wands/ModularWandItemNeoforge.java
@@ -1,10 +1,10 @@
 package com.verdantartifice.primalmagick.common.items.wands;
 
-import com.verdantartifice.primalmagick.common.wands.IWandComponent;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.context.UseOnContext;
+import org.jetbrains.annotations.NotNull;
 
 public class ModularWandItemNeoforge extends ModularWandItem {
     public ModularWandItemNeoforge(Item.Properties properties) {
@@ -12,24 +12,19 @@ public class ModularWandItemNeoforge extends ModularWandItem {
     }
 
     @Override
-    public InteractionResult onItemUseFirst(ItemStack stack, UseOnContext context) {
+    @NotNull
+    public InteractionResult onItemUseFirst(@NotNull ItemStack stack, @NotNull UseOnContext context) {
         return this.onWandUseFirst(stack, context);
     }
 
     @Override
-    public boolean shouldCauseReequipAnimation(ItemStack oldStack, ItemStack newStack, boolean slotChanged) {
+    public boolean shouldCauseReequipAnimation(@NotNull ItemStack oldStack, @NotNull ItemStack newStack, boolean slotChanged) {
         return false;
     }
 
     @Override
-    public boolean canContinueUsing(ItemStack oldStack, ItemStack newStack) {
+    public boolean canContinueUsing(@NotNull ItemStack oldStack, @NotNull ItemStack newStack) {
         // Don't break wand interaction just because the stack NBT changes
         return true;
-    }
-
-    @Override
-    public int getEnchantmentValue(ItemStack stack) {
-        // The enchantability of a wand is determined by its components
-        return this.getComponents(stack).stream().mapToInt(IWandComponent::getEnchantability).sum();
     }
 }